Add files via upload
The first image is the start menu that shows the options the player has when they start the game. The second picture shows the game interface. The rules are the same as in a normal Pacman game, however if one of the pacman gets the tablet or colored circle, they can eat the other Pacman which results in a win for that player and a loss for the other pacman. The other way to win is whoever eats the most ghosts.